requirements.txt
python-telegram-bot==20.6
PyPDF2==3.0.1
